Themes: Set a value for osd.window-switcher.width
This setting does not exist in Openbox, but does in labwc. In labwc,
the "*.width: 0" entry naturally causes the window switcher width to
also be set to 0. As this is written, a value of 0 causes labwc to
segfault. I expect this to be fixed soon (I have filed a report). Even
so, it makes sense to set an actual value for this setting because a
window switcher width of 0 doesn't really make sense in any case.

The width of 600 was chosen because it is the default for labwc.