Sfwbar: Switch workspaces by scrolling on the bar.
In Openbox I usually switch workspaces by using the scroll wheel on
the desktop. This naturally does not work if there is a maximized
window. However, one of the nice coincidences of my Openbox setup is
that scrolling on the tint2 bar also switches workspaces.

This is not the case with labwc/sfwbar and I have had trouble
replicating this behavior since sfwbar does not have any actions to
switch workspaces. However, I can use the wtype command to send my
"switch to next/previous" keybind, which will be picked up by labwc.

This has to be applied to each widget in sfwbar to make it apply
across the entire bar. This is ugly, but what can you do.