Labwc/Sfwbar: Right click on taskbar blank area to show root menu.
This is another openbox/tint2 behavior that can be hacked into
labwc/sfwbar via wtype. labwc needs a new keybind for showing the root
menu so that wtype can trigger that action. I have set it up to use
super + numpad 0 since that is a keybind that I am unlikely to ever
need for some other purpose. Further wtype-specific keybinds can use
numpad 1, numpad 2, etc. in the case that I ever need them.

Currently, labwc has no way to prevent the root menu from overflowing
off the screen if there are too many items, which is an issue if the
root menu is triggered near the edge of the screen (like where the
taskbar is). Also, sfwbar itself is drawn over the top of the root
menu. These are things that may fix themselves in time.