Bash: Rework shell functions.
I have several "emacs commands" that are functions or aliases that
call emacsclient in the terminal. I have rewritten these so that when
they are called from the Emacs terminal (Eat), they run the Emacs
commands directly instead of opening emacsclient within the terminal.
If I am not using Eat, the functions behave as before.

I have also modified the 'dired' and 'magit' commands to be more
robust, added a new 'man' function, sorted the functions by name, and
added a note clarifying what kind of functions belong in .bashrc.