Sfwbar: Explicitly release modifier keys in wtype commands.
Wtype is supposed to release all modifier keys when it exits, but I
was getting some weird behavior, with the desktop switcher OSD staying
on screen indefinitely (it's supposed to disappear when the modifier
is released).

Now the OSD goes so fast as to be almost invisible, but I think that's
strictly better than not going away at all.